+++ src/modules/domSelections.js
@@ -9,3 +9,7 @@ export const scoreDom = document.querySelector(
 export const submitButton = document.querySelector(
   '.landing__section-addScore-button',
 );
+
+export const refreshButton = document.querySelector(
+  '.landing__section-recentScores-button',
+);
