--- src/modules/env.js
@@ -1,1 +1,0 @@
-export const baseUrl = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/';
